requirements.txt
flask
google-generativeai
gunicorn
diffusers==0.22.1
transformers
scipy 
ftfy 
accelerate